重构代码以使用新命名和类型
重构多个文件以使用新的命名和类型，包括将 `ApiKey` 改为 `UserApiKey` 或 `Secret`，将 `ApiUsage2` 改为 `UserApiUsage`，并更新相关的数据库查询和方法参数。此外，添加了新的类 `ClientInfo`、`MessageRequest` 和 `Tokenizer`，并更新了现有类的属性类型和导航属性。

+++ src/BE/Controllers/Admin/AdminModels/Dtos/UpdateModelRequest.cs
@@ -36,7 +36,7 @@ public record UpdateModelRequest
 
     public void ApplyTo(Model cm, ChatsDB db)
     {
-        int? modelReferenceId = db.ModelReferences
+        short? modelReferenceId = db.ModelReferences
             .Where(x => x.Name == ModelReferenceName && x.ProviderId == db.Models.Where(x => x.Id == ModelKeyId).Select(x => x.Id).First())
             .Select(x => x.Id)
             .FirstOrDefault();
